requirements.txt
requests==2.28.2
retry2==0.9.5
